Avoid resource_control drift due to whitespace in queries
As reported in hashicorp/terraform#23928, reading a string from a file
and comparing it to the terraform state may produce a drift due to
inconsistencies (ie. Linux vs Windows line endings). This PR suppresses
any drift caused by differences between the desired state and the stored
state, by unmarshalling the JSONs and comparing the resulting objects.